# specs/technical/management-api.openapi.yaml
openapi: "3.1.0"
...
info:
  title: skoed Management API
  version: "2.0.0"
  x-cluster-write-semantics: |
    As of M2, every mutating endpoint under /api/v1 (blocklists, allowlist,
    local-dns, settings, config/import, auth) is cluster-aware:

      * If the receiving node is the Raft leader, the write is applied through
        the FSM, replicated, and the response is returned once it commits.
      * If the receiving node is a follower, the request is transparently
        forwarded to the current leader over the same HTTP API. The admin
        sees no difference; only the latency is slightly higher.
      * If no leader is currently elected (e.g. minority partition), the call
        returns HTTP 503 with a LeaderRedirect body whose `leader_id` is empty.

    Read endpoints (GET) are always served locally from bbolt without
    forwarding, so they remain available during leader transitions.

    See specs/technical/raft-fsm.md for the FSM command set and write path.
  description: >
    REST API for managing a skoed node. All endpoints require HTTP Basic
    authentication except POST /api/v1/auth/setup (first-run only).
    DNS port 53 is never gated by this API's authentication.
servers:
  - url: http://{host}:{port}
    variables:
      host:
        default: localhost
      port:
        default: "8080"
security:
  - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    LeaderRedirect:
      type: object
      required: [error, leader_address]
      properties:
        error:
          type: string
          example: "not the leader"
        leader_id:
          type: string
        leader_address:
          type: string
          example: "192.168.1.10:8080"
...
paths:
...
  /api/v1/cluster/leadership/transfer:
    post:
      summary: Transfer leadership to a specific node
      description: >
        Triggers a Raft leadership transfer to the named follower. The call
        returns once the transfer has completed or timed out.
      operationId: transferLeadership
      x-tsid: TS-ClusterApi
      x-fsid-links:
        - FS-LeaderFailoverManualTransfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_node_id]
              properties:
                target_node_id:
                  type: string
      responses:
        "204":
          description: Leadership transferred
        "409":
          description: Not the leader OR target not in cluster
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderRedirect"
        "504":
          description: Transfer did not complete within the deadline
...